move test to healthchecks
