+++ SingularityService/src/test/java/com/hubspot/singularity/scheduler/SingularityHealthchecksTest.java
@@ -1,5 +1,7 @@
 package com.hubspot.singularity.scheduler;
 
+import java.util.Arrays;
+import java.util.Collections;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
@@ -8,6 +10,7 @@ import org.junit.Assert;
 import org.junit.Test;
 
 import com.google.common.base.Optional;
+import com.hubspot.mesos.Resources;
 import com.hubspot.singularity.DeployState;
 import com.hubspot.singularity.SingularityDeploy;
 import com.hubspot.singularity.SingularityDeployBuilder;
@@ -266,6 +269,41 @@ public class SingularityHealthchecksTest extends SingularitySchedulerTestBase {
     deployChecker.checkDeploys();
 
     Assert.assertEquals(DeployState.SUCCEEDED, deployManager.getDeployResult(requestId, deployId).get().getDeployState());
+  }
+
+  @Test
+  public void testPortIndices() {
+    configuration.setNewTaskCheckerBaseDelaySeconds(0);
+    configuration.setHealthcheckIntervalSeconds(0);
+    configuration.setDeployHealthyBySeconds(0);
+    configuration.setKillAfterTasksDoNotRunDefaultSeconds(1);
+    configuration.setHealthcheckMaxRetries(Optional.of(0));
+
+    initRequest();
+    firstDeploy = initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), firstDeployId)
+      .setCommand(Optional.of("sleep 100"))
+      .setHealthcheckUri(Optional.of("http://uri"))
+      .setResources(Optional.of(new Resources(1, 64, 3, 0)))
+      .setHealthcheckPortIndex(Optional.of(1)));
+
+    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
+    scheduler.drainPendingQueue(stateCacheProvider.get());
+
+    String[] portRange = {"80:82"};
+    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String> absent(), Collections.<String, String>emptyMap(), portRange)));
+
+    SingularityTaskId firstTaskId = taskManager.getActiveTaskIdsForRequest(requestId).get(0);
+
+    SingularityTask firstTask = taskManager.getTask(firstTaskId).get();
+    statusUpdate(firstTask, TaskState.TASK_RUNNING);
+
+    newTaskChecker.enqueueNewTaskCheck(firstTask, requestManager.getRequest(requestId), healthchecker);
+
+    finishNewTaskChecks();
+    finishHealthchecks();
+    finishNewTaskChecksAndCleanup();
+
+    Assert.assertTrue(taskManager.getLastHealthcheck(firstTask.getTaskId()).get().toString().contains("host1:81"));
   }
 
 }
