+++ SingularityService/src/test/java/com/hubspot/singularity/scheduler/SingularitySchedulerTest.java
@@ -1406,41 +1406,6 @@ public class SingularitySchedulerTest extends SingularitySchedulerTestBase {
   }
 
   @Test
-  public void testPortIndices() {
-    configuration.setNewTaskCheckerBaseDelaySeconds(0);
-    configuration.setHealthcheckIntervalSeconds(0);
-    configuration.setDeployHealthyBySeconds(0);
-    configuration.setKillAfterTasksDoNotRunDefaultSeconds(1);
-    configuration.setHealthcheckMaxRetries(Optional.of(0));
-
-    initRequest();
-    firstDeploy = initAndFinishDeploy(request, new SingularityDeployBuilder(request.getId(), firstDeployId)
-      .setCommand(Optional.of("sleep 100"))
-      .setHealthcheckUri(Optional.of("http://uri"))
-      .setResources(Optional.of(new Resources(1, 64, 3, 0)))
-      .setHealthcheckPortIndex(Optional.of(1)));
-
-    requestResource.postRequest(request.toBuilder().setInstances(Optional.of(2)).build());
-    scheduler.drainPendingQueue(stateCacheProvider.get());
-
-    String[] portRange = {"80:82"};
-    sms.resourceOffers(driver, Arrays.asList(createOffer(20, 20000, "slave1", "host1", Optional.<String> absent(), Collections.<String, String>emptyMap(), portRange)));
-
-    SingularityTaskId firstTaskId = taskManager.getActiveTaskIdsForRequest(requestId).get(0);
-
-    SingularityTask firstTask = taskManager.getTask(firstTaskId).get();
-    statusUpdate(firstTask, TaskState.TASK_RUNNING);
-
-    newTaskChecker.enqueueNewTaskCheck(firstTask, requestManager.getRequest(requestId), healthchecker);
-
-    finishNewTaskChecks();
-    finishHealthchecks();
-    finishNewTaskChecksAndCleanup();
-
-    Assert.assertTrue(taskManager.getLastHealthcheck(firstTask.getTaskId()).get().toString().contains("host1:81"));
-  }
-
-  @Test
   public void testQueueMultipleOneOffs() {
     SingularityRequestBuilder bldr = new SingularityRequestBuilder(requestId, RequestType.ON_DEMAND);
     requestResource.postRequest(bldr.build());
